chore: moved fields layout related code to crm_fields_layout.py
