+++ crm/api/doc.py
@@ -157,62 +157,6 @@ def get_group_by_fields(doctype: str):
 	return fields
 
 
-@frappe.whitelist()
-def get_fields_layout(doctype: str, type: str):
-	sections = []
-	if frappe.db.exists("CRM Fields Layout", {"dt": doctype, "type": type}):
-		layout = frappe.get_doc("CRM Fields Layout", {"dt": doctype, "type": type})
-	else:
-		return []
-
-	if layout.layout:
-		sections = json.loads(layout.layout)
-
-	allowed_fields = []
-	for section in sections:
-		if not section.get("fields"):
-			continue
-		allowed_fields.extend(section.get("fields"))
-
-	fields = frappe.get_meta(doctype).fields
-	fields = [field for field in fields if field.fieldname in allowed_fields]
-
-	for section in sections:
-		for field in section.get("fields") if section.get("fields") else []:
-			field = next((f for f in fields if f.fieldname == field), None)
-			if field:
-				if field.fieldtype == "Select":
-					field.options = field.options.split("\n")
-					field.options = [{"label": _(option), "value": option} for option in field.options]
-					field.options.insert(0, {"label": "", "value": ""})
-				field = {
-					"label": _(field.label),
-					"name": field.fieldname,
-					"type": field.fieldtype,
-					"options": field.options,
-					"mandatory": field.reqd,
-				}
-				section["fields"][section.get("fields").index(field["name"])] = field
-
-	return sections or []
-
-
-@frappe.whitelist()
-def save_fields_layout(doctype: str, type: str, layout: str):
-	if frappe.db.exists("CRM Fields Layout", {"dt": doctype, "type": type}):
-		doc = frappe.get_doc("CRM Fields Layout", {"dt": doctype, "type": type})
-	else:
-		doc = frappe.new_doc("CRM Fields Layout")
-
-	doc.update({
-		"dt": doctype,
-		"type": type,
-		"layout": layout,
-	})
-	doc.save(ignore_permissions=True)
-
-	return doc.layout
-
 def get_doctype_fields_meta(DocField, doctype, allowed_fieldtypes, restricted_fields):
 	parent = "parent" if DocField._table_name == "tabDocField" else "dt"
 	return (
