+++ frontend/src/components/Modals/DealModal.vue
@@ -77,7 +77,7 @@ const chooseExistingContact = ref(false)
 const chooseExistingOrganization = ref(false)
 
 const sections = createResource({
-  url: 'crm.api.doc.get_fields_layout',
+  url: 'crm.fcrm.doctype.crm_fields_layout.crm_fields_layout.get_fields_layout',
   cache: ['quickEntryFields', 'CRM Deal'],
   params: { doctype: 'CRM Deal', type: 'Quick Entry'},
   auto: true,
